+++ openrouteservice-api-tests/src/test/java/heigit/ors/services/isochrones/ResultTest.java
@@ -1,22 +1,24 @@
 /*
- *  Licensed to GIScience Research Group, Heidelberg University (GIScience)
  *
- *   	 http://www.giscience.uni-hd.de
- *   	 http://www.heigit.org
+ *  *  Licensed to GIScience Research Group, Heidelberg University (GIScience)
+ *  *
+ *  *   	 http://www.giscience.uni-hd.de
+ *  *   	 http://www.heigit.org
+ *  *
+ *  *  under one or more contributor license agreements. See the NOTICE file
+ *  *  distributed with this work for additional information regarding copyright
+ *  *  ownership. The GIScience licenses this file to you under the Apache License,
+ *  *  Version 2.0 (the "License"); you may not use this file except in compliance
+ *  *  with the License. You may obtain a copy of the License at
+ *  *
+ *  *       http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  *  Unless required by applicable law or agreed to in writing, software
+ *  *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  *  See the License for the specific language governing permissions and
+ *  *  limitations under the License.
  *
- *  under one or more contributor license agreements. See the NOTICE file 
- *  distributed with this work for additional information regarding copyright 
- *  ownership. The GIScience licenses this file to you under the Apache License, 
- *  Version 2.0 (the "License"); you may not use this file except in compliance 
- *  with the License. You may obtain a copy of the License at
- * 
- *       http://www.apache.org/licenses/LICENSE-2.0
- * 
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
  */
 package heigit.ors.services.isochrones;
 
@@ -94,7 +96,7 @@ public class ResultTest extends ServiceTest {
 				.get(getEndPointName())
 				.then()
 				.statusCode(500)
-				.body("error.code", is(399));
+				.body("error.code", is(IsochronesErrorCodes.UNKNOWN));
 	}
 
 	@Test