+++ openrouteservice/src/main/java/heigit/ors/routing/graphhopper/extensions/edgefilters/AvoidBordersEdgeFilter.java
@@ -1,3 +1,23 @@
+/*
+ *  Licensed to GIScience Research Group, Heidelberg University (GIScience)
+ *
+ *   http://www.giscience.uni-hd.de
+ *   http://www.heigit.org
+ *
+ *  under one or more contributor license agreements. See the NOTICE file
+ *  distributed with this work for additional information regarding copyright
+ *  ownership. The GIScience licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except in compliance
+ *  with the License. You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 package heigit.ors.routing.graphhopper.extensions.edgefilters;
 
 import com.graphhopper.routing.util.EdgeFilter;
@@ -49,6 +69,12 @@ public class AvoidBordersEdgeFilter implements EdgeFilter {
         this._encoder = encoder;
     }
 
+    /**
+     * Initialise the edge filter object based on the type of borders to filter
+     *
+     * @param searchParams      The search parameters passed into the request
+     * @param extBorders        The extended borders graph storage to use
+     */
     private void init(RouteSearchParameters searchParams, BordersGraphStorage extBorders) {
         // Init the graph storage
         if(extBorders != null) {
@@ -68,6 +94,13 @@ public class AvoidBordersEdgeFilter implements EdgeFilter {
         }
     }
 
+    /**
+     * Determine whether the edge should be accepted for processing or reject. Depending on whether the request was to
+     * not cross any border or not cross controlled borders determines the type of border to reject.
+     *
+     * @param iter      An iterator to the edges that need to be filtered
+     * @return
+     */
     @Override
     public final boolean accept(EdgeIteratorState iter) {