+++ CHANGELOG.md
@@ -5,6 +5,54 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+-
+
+### Fixed
+- Fixed problem with avoid polygons excluding ways that should have been accepted (Issue #95)
+
+### Changed
+-
+
+### Depricated
+-
+
+## [4.5.1] - 2018-03-20
+### Fixed
+- Fixed the parameters being the wrong way around in isochrones request when maximum range has been exceeded (Issue #126)
+- Fixed the coordinate precision in the geojson export from 4 to 6 decimals (Issue #136)
+- Fixed the instructions='false' error when exporting as geojson (Issue #138)
+- Fixed missing summary in the geojson output (Issue #139)
+- Fixed error when a high exit number for a roundabout is used in instructions (Issue #145)
+
+### Changed
+- Updated error response code for routing when no route can be found between locations (Issue #144)
+- Updated logging so that stack traces are only output when debug logging is enabled (Issue #148)
+- Updated the error response for geocding when no address found (Issue #134)
+
+
+## [4.5] - 2018-02-27
+### Added
+- Functionality has been added to restrict routes so that they do not cross all borders, controlled borders, or the borders of specific countries (Issue #41)
+- Added GeoJson export for routing exports (Issue #54)
+- Added global export class to combine all exports there (Issue #123)
+- Option to specify maximum locations for matrix request when using non-standard weightings (Issue #94)
+
+### Fixed
+- Fix exception when roundabout exit is not correctly found (Issue #89)
+- Option to specify maximum locations for matrix request when using non-standard weightings (Issue #94)
+- Geocoder now returns a 404 response if no address is found for reverse geocoding (Issue #113)
+- Fixed error codes (Issue #109)
+- Correct querying of population statistics data for isochrones (Issue #106)
+
+### Changed
+- RoutingProfile was changed to make sure whenever pop_total or pop_area is queried, both are present in the attributes (Issue #106)
+- Response with a detour factor now uses "detourfactor" rather than "detour_factor" (Issue #61)
+- Changed the gpx export to the new global export processor (Issue #123)
+
+### Deprecated
+- getStatisticsOld | Connected to the old statistics library (Issue #106)
+- geometryToWKB | Connected to the old statistics library (Issue #106)
 
 ## [4.4.2] - 2018-01-31
 ### Added
@@ -52,4 +100,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - 
 
 ### Deprecated
--
+- 
+